Extract scheduler logic
